Move some function to util files
