+++ flamechart/src/utils/filter-data.ts
@@ -1,0 +1,66 @@
+// Copyright 2025 The Perses Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { StackTrace } from '@perses-dev/core';
+
+/**
+ * Filter the global stacktrace by a function ID to focus on that function and display its corresponding flame chart
+ */
+export function filterJson(json: StackTrace, id?: number): StackTrace {
+  if (id === null) {
+    return json;
+  }
+
+  const recur = (item: StackTrace, id?: number): StackTrace | undefined => {
+    if (item.id === id) {
+      return item;
+    }
+
+    for (const child of item.children || []) {
+      const temp = recur(child, id);
+      if (temp) {
+        item.children = [temp];
+
+        // change the parents' values (todo : verify this)
+        item.start = temp.start;
+        item.end = temp.end;
+        // item.self = temp.self;
+        // item.total = temp.total;
+
+        return item;
+      }
+    }
+  };
+
+  return recur(json, id) || json;
+}
+
+/**
+ * Calculate the maximum depth of the stack trace
+ */
+export function heightOfJson(json: StackTrace): number {
+  const recur = (item: StackTrace): number => {
+    if ((item.children || []).length === 0) {
+      return item.level;
+    }
+
+    let maxLevel = item.level;
+    for (const child of item.children!) {
+      const tempLevel = recur(child);
+      maxLevel = Math.max(maxLevel, tempLevel);
+    }
+    return maxLevel;
+  };
+
+  return recur(json);
+}
